+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - BREAKING FEATURE: add `max_output_steps` option to `gen_path.line()` and `gen_path.angle()` to limit the number of steps the return. Default is 1000 to improve performance on large cursor jumps which also is set for `config.cursor.path`.
 
+## mini.hipatterns
+
+- BREAKING FEATURE: make `MiniHipatterns{Fixme,Hack,Todo,Note}` highlight groups by default be reverse and bold variant of `Diagnostic{Error,Warn,Info,Hint}` group instead of directly link to them. This ensures better visibility for color schemes which don't have explicit 'mini.hipatterns' support.
+
 ## mini.hues
 
 - FEATURE: add `'lowmedium'` and `'mediumhigh'` saturation levels.